# Use runtime-selected implementation for stage 1 on demand