--- a/include/simdjson/generic/ondemand/parser-inl.h
+++ b/include/simdjson/generic/ondemand/parser-inl.h
@@ -10,8 +10,12 @@ simdjson_warn_unused simdjson_really_inline error_code parser::allocate(size_t n
   _max_depth = 0;
   size_t string_capacity = SIMDJSON_ROUNDUP_N(5 * new_capacity / 3 + SIMDJSON_PADDING, 64);
   string_buf.reset(new (std::nothrow) uint8_t[string_capacity]);
-  SIMDJSON_TRY( dom_parser.set_capacity(new_capacity) );
-  SIMDJSON_TRY( dom_parser.set_max_depth(DEFAULT_MAX_DEPTH) );
+  if (dom_parser) {
+    SIMDJSON_TRY( dom_parser->set_capacity(new_capacity) );
+    SIMDJSON_TRY( dom_parser->set_max_depth(new_max_depth) );
+  } else {
+    SIMDJSON_TRY( simdjson::active_implementation->create_dom_parser_implementation(new_capacity, new_max_depth, dom_parser) );
+  }
   _capacity = new_capacity;
   _max_depth = new_max_depth;
   return SUCCESS;
@@ -24,8 +28,8 @@ simdjson_warn_unused simdjson_really_inline simdjson_result<document> parser::it
   }
 
   // Run stage 1.
-  SIMDJSON_TRY( dom_parser.stage1((const uint8_t *)buf.data(), buf.size(), false) );
-  return document::start(this);
+  SIMDJSON_TRY( dom_parser->stage1((const uint8_t *)buf.data(), buf.size(), false) );
+  return document::start({ (const uint8_t *)buf.data(), this });
 }
 
 simdjson_warn_unused simdjson_really_inline simdjson_result<document> parser::iterate(const simdjson_result<padded_string> &result) & noexcept {
@@ -42,8 +46,8 @@ simdjson_warn_unused simdjson_really_inline simdjson_result<json_iterator> parse
   }
 
   // Run stage 1.
-  SIMDJSON_TRY( dom_parser.stage1((const uint8_t *)buf.data(), buf.size(), false) );
-  return json_iterator(this);
+  SIMDJSON_TRY( dom_parser->stage1((const uint8_t *)buf.data(), buf.size(), false) );
+  return json_iterator((const uint8_t *)buf.data(), this);
 }
 
 } // namespace ondemand