Fixes #1358.

It's clearly better. In fact, it's unequivocally better, in all circumstances. Oddly, in many cases (most of the benchmarks, in fact) the *fallback* frontend beats the haswell frontend. Not sure what to think of that. Maybe jitters. I ran all of them twice to check (all 4 once, then all 4 again).

| Branch | Host | Compiler | cmake | Stage 1 | On Demand | partial_tweets | large_random | large_random_unordered | kostya | distinct_user_id | find_tweet |
|---|---|---|---|---|---|---|---|---|---|
| PR | Skylake | Clang 10 | CXX_FLAGS=-march=native | haswell | haswell | **4.10 GB/s** | 0.69 GB/s | 0.70 GB/s | 2.25 GB/s | 3.97 GB/s | 5.80 GB/s |
| PR | Skylake | Clang 10 | | haswell | fallback | 3.55 GB/s | **0.70 GB/s** | **0.71 GB/s** | **2.27 GB/s** | **4.23 GB/s** | **5.83 GB/s** |
| Before | Skylake | Clang 10 | CXX_FLAGS=-march=native | haswell | haswell |  3.94 GB/s | 0.69 GB/s | 0.69 GB/s | 2.15 GB/s | 4.07 GB/s | 5.64 GB/s |
| Before | Skylake | Clang 10 | | fallback | 0.51 GB/s | 0.42 GB/s | 0.42 GB/s | 0.69 GB/s | 0.53 GB/s | 0.55 GB/s |