@jkeiser 2021-01-07
This is super exciting. -march=native is not something people are likely to run when writing benchmarks and testing things out (while -O3 is). Plus, people who want to run their code on multiple hosts won't *want* -march=native: it's hard to argue that it's universal, because it's not.