@jkeiser 2021-01-07
There are some API changes that might not be terrible at this stage, but this is the last substantial change *necessary* to release, IMO. The rest is writeup.

The changes that might still be nice:

- Remove the need to write `simdjson::builtin`. Alias simdjson::builtin::ondemand to simdjson::ondemand.
- Allow scalar values to be used out of order. Still can't do it for arrays, but scalars are more likely to be (maybe accidentally) used like that anyway. While this is another betrayal of the "forward only" mentality, making the library as unsurprising as possible seems important for adoption. Array and hash iteration still needs to be forward only, because of the single iterator (which is an important concept for performance: let the user do the iteration work for us during the natural course of iteration).