@lemire 2021-01-07
I was planning to work on this today, but now I will get to review your excellent code instead.

The numbers are much better than I expected. In fact, it appears to largely solve my concerns.

This is a brilliant move, John. Brilliant.