@jkeiser 2021-01-11
This one was based against ondemand-sajson, only because I wanted to test it against all these other implementations. Rebasing against master and merging.